requirements.txt
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.119.1
griffe==1.14.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
jiter==0.11.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mcp==1.18.0
multidict==6.7.0
openai==2.6.0
openai-agents==0.4.0
packaging==25.0
postgrest==2.22.0
propcache==0.4.1
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
PyJWT==2.10.1
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
realtime==2.22.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.27.1
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.48.0
storage3==2.22.0
StrEnum==0.4.15
supabase==2.22.0
supabase-auth==2.22.0
supabase-functions==2.22.0
tqdm==4.67.1
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
